+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [10.0.0] 2024-11-12
+
+### Added
+
+* Add optional storage for complete prompt and response using dynamodb and media manager.
+
+### Removed
+
+* Eliminated old openai code
 
 ### Fixed